# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Nexspence API
  description: |
    Nexspence REST API — two surfaces:
    - `/service/rest/v1/` — **Nexus v1 compatible** (drop-in replacement)
    - `/api/v1/` — **Native Nexspence API** (extended features + migration)

    ## Authentication
    All endpoints (except anonymous-allowed) require one of:
    - **HTTP Basic Auth** — `Authorization: Basic base64(user:pass)`
    - **Bearer token** — `Authorization: Bearer <api-token>`
  version: 1.0.0
  contact:
    name: Nexspence
servers:
  - url: http://localhost:8080
    description: Local development

# ================================================================
# SECURITY SCHEMES
# ================================================================
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer

  # ── Common schemas ──────────────────────────────────────────
...
security:
  - basicAuth: []
  - bearerAuth: []

# ================================================================
# PATHS
# ================================================================
paths:
...
  /service/rest/v1/security/users/{userId}/change-password:
    put:
      tags: [Security]
      summary: Reset a user's password (admin)
      description: >
        Admin reset: no `oldPassword` is required. Only local accounts can be
        reset — ldap/oidc/saml credentials belong to the identity provider and
        are refused with 403. The reset bumps the target's JWT cutoff, ending
        their existing browser sessions. Users changing their own password use
        `/api/v1/me/change-password`, which always verifies the current one.
      operationId: changePassword
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword: { type: string, description: 'At least `auth.password_min_length` characters' }
      responses:
        '204': { description: Changed }
        '400': { description: The new password is below auth.password_min_length }
        '403': { description: The account's password is managed by an identity provider }
        '404': { description: No such user }

  # ── Security — Roles ─────────────────────────────────────────
...